Applied PoW refactoring to Dogecoin specific functions
- Moved GetNextWorkRequired(), CheckProofOfWork() and ComputeMinWork()
  out of main.cpp, into pow.cpp
- specified CChainParams generic nTargetTimespan and nTargetSpacing
- created Params().GetDigiShieldForkBlock() which returns the block where
  DigiShield is activated
- created Params().DigiShieldTargetTimespan() and DigiShieldInterval(),
  which return the post-DigiShield equivalents of .TargetTimespan() and
  .Interval()
- refactored the PoW functions to read from Params() instead of locally
  declared consts.
