+++ src/main.cpp
@@ -16,6 +16,7 @@
 #include "init.h"
 #include "auxpow.h"
 #include "net.h"
+#include "pow.h"
 #include "txdb.h"
 #include "txmempool.h"
 #include "ui_interface.h"
@@ -1323,197 +1324,6 @@ int64_t GetBlockValue(int nHeight, int64_t nFees, uint256 prevHash)
     }
 
     return nSubsidy + nFees;
-}
-
-// New Difficulty adjustement and reward scheme by /u/lleti, rog1121, and DigiByte (DigiShield Developers).
-static const int64_t nTargetTimespan = 4 * 60 * 60; // Dogecoin: every 4 hours
-static const int64_t nTargetTimespanNEW = 60 ; // Dogecoin: every 1 minute
-static const int64_t nTargetSpacing = 60; // Dogecoin: 1 minute
-static const int64_t nInterval = nTargetTimespan / nTargetSpacing;
-
-static const int64_t nDiffChangeTarget = 145000; // Patch effective @ block 145000
-
-//
-// minimum amount of work that could possibly be required nTime after
-// minimum work required was nBase
-//
-unsigned int ComputeMinWork(unsigned int nBase, int64_t nTime)
-{
-    const uint256 &bnLimit = Params().ProofOfWorkLimit();
-    // Testnet has min-difficulty blocks
-    // after nTargetSpacing*2 time between blocks:
-    if (Params().AllowMinDifficultyBlocks() && nTime > nTargetSpacing*2)
-        return bnLimit.GetCompact();
-
-    uint256 bnResult;
-    bnResult.SetCompact(nBase);
-    while (nTime > 0 && bnResult < bnLimit)
-    {
-        if(chainActive.Height()+1<nDiffChangeTarget){
-            // Maximum 400% adjustment...
-            bnResult *= 4;
-            // ... in best-case exactly 4-times-normal target time
-            nTime -= nTargetTimespan*4;
-        } else {
-            // Maximum 10% adjustment...
-            bnResult = (bnResult * 110) / 100;
-            // ... in best-case exactly 4-times-normal target time
-            nTime -= nTargetTimespanNEW*4;
-        }
-    }
-    if (bnResult > bnLimit)
-        bnResult = bnLimit;
-    return bnResult.GetCompact();
-}
-
-// Determine if the for the given block, a min difficulty setting applies
-bool AllowMinDifficultyForBlock(const CBlockIndex* pindexLast, const CBlockHeader *pblock)
-{
-    // check if the chain allows minimum difficulty blocks
-    if (!Params().AllowMinDifficultyBlocks())
-        return false;
-
-    // check if we allow minimum difficulty at this block-height
-    if (pindexLast->nHeight < Params().GetMinDifficultyAllowedStartBlock())
-        return false;
-
-    // Allow for a minimum block time if the elapsed time > 2*nTargetSpacing
-    return (pblock->nTime > pindexLast->nTime + nTargetSpacing*2);
-}
-
-unsigned int GetNextWorkRequired(const CBlockIndex* pindexLast, const CBlockHeader *pblock)
-{
-    unsigned int nProofOfWorkLimit = Params().ProofOfWorkLimit().GetCompact();
-
-    int nHeight = pindexLast->nHeight + 1;
-    bool fNewDifficultyProtocol = (nHeight >= nDiffChangeTarget);
-    
-    int64_t retargetTimespan = nTargetTimespan;
-    int64_t retargetInterval = nInterval;
-    
-    if (fNewDifficultyProtocol) {
-        retargetInterval = nTargetTimespanNEW / nTargetSpacing;
-        retargetTimespan = nTargetTimespanNEW;
-    }
-    
-    // Genesis block
-    if (pindexLast == NULL)
-        return nProofOfWorkLimit;
-
-    if (AllowMinDifficultyForBlock(pindexLast, pblock))
-    {
-        // Special difficulty rule for testnet:
-        // If the new block's timestamp is more than 2* nTargetSpacing minutes
-        // then allow mining of a min-difficulty block.
-        return nProofOfWorkLimit;
-    }
-
-    // Only change once per interval
-    if ((pindexLast->nHeight+1) % retargetInterval != 0)
-    {
-        // kept for testnet legacy pre-block 160000
-        if (Params().AllowMinDifficultyBlocks())
-        {
-            if (pblock->nTime > pindexLast->nTime + nTargetSpacing*2)
-            {
-                // Special difficulty rule for testnet:
-                // If the new block's timestamp is more than 2* nTargetSpacing minutes
-                // then allow mining of a min-difficulty block.
-                return nProofOfWorkLimit;
-            } else {
-                // Return the last non-special-min-difficulty-rules-block
-                const CBlockIndex* pindex = pindexLast;
-                while (pindex->pprev && pindex->nHeight % retargetInterval != 0 && pindex->nBits == nProofOfWorkLimit)
-                    pindex = pindex->pprev;
-                return pindex->nBits;
-            }
-        }
-        return pindexLast->nBits;
-    }
-
-    // Dogecoin: This fixes an issue where a 51% attack can change difficulty at will.
-    // Go back the full period unless it's the first retarget after genesis. Code courtesy of Art Forz
-    int blockstogoback = retargetInterval-1;
-    if ((pindexLast->nHeight+1) != retargetInterval)
-        blockstogoback = retargetInterval;
-
-    // Go back by what we want to be 14 days worth of blocks
-    const CBlockIndex* pindexFirst = pindexLast;
-    for (int i = 0; pindexFirst && i < blockstogoback; i++)
-        pindexFirst = pindexFirst->pprev;
-    assert(pindexFirst);
-
-    // Limit adjustment step
-    int64_t nActualTimespan = pindexLast->GetBlockTime() - pindexFirst->GetBlockTime();
-    int64_t nModulatedTimespan = nActualTimespan;
-    
-    if (fNewDifficultyProtocol) //DigiShield implementation - thanks to RealSolid & WDC for this code
-    {
-        // amplitude filter - thanks to daft27 for this code
-        nModulatedTimespan = retargetTimespan + (nModulatedTimespan - retargetTimespan)/8;
-
-        if (nModulatedTimespan < (retargetTimespan - (retargetTimespan/4)) ) nModulatedTimespan = (retargetTimespan - (retargetTimespan/4));
-        if (nModulatedTimespan > (retargetTimespan + (retargetTimespan/2)) ) nModulatedTimespan = (retargetTimespan + (retargetTimespan/2));
-    }
-    else if (pindexLast->nHeight+1 > 10000) {
-        if (nModulatedTimespan < nTargetTimespan/4)
-            nModulatedTimespan = nTargetTimespan/4;
-        if (nModulatedTimespan > nTargetTimespan*4)
-            nModulatedTimespan = nTargetTimespan*4;
-    }
-    else if (pindexLast->nHeight+1 > 5000)
-    {
-        if (nModulatedTimespan < nTargetTimespan/8)
-            nModulatedTimespan = nTargetTimespan/8;
-        if (nModulatedTimespan > nTargetTimespan*4)
-            nModulatedTimespan = nTargetTimespan*4;
-    }
-    else
-    {
-        if (nModulatedTimespan < nTargetTimespan/16)
-            nModulatedTimespan = nTargetTimespan/16;
-        if (nModulatedTimespan > nTargetTimespan*4)
-            nModulatedTimespan = nTargetTimespan*4;
-    }
-
-    // Retarget
-    uint256 bnNew;
-    uint256 bnOld;
-    bnNew.SetCompact(pindexLast->nBits);
-    bnOld = bnNew;
-    bnNew *= nModulatedTimespan;
-    bnNew /= retargetTimespan;
-
-    if (bnNew > Params().ProofOfWorkLimit())
-        bnNew = Params().ProofOfWorkLimit();
-
-    unsigned int nNewBits = bnNew.GetCompact();
-
-    /// debug print
-    LogPrintf("GetNextWorkRequired RETARGET\n");
-    LogPrintf("nTargetTimespan = %d    nActualTimespan = %d    nModulated = %d\n", nTargetTimespan, nActualTimespan, nModulatedTimespan);
-    LogPrintf("Before: %08x  %s\n", pindexLast->nBits, bnOld.ToString());
-    LogPrintf("After:  %08x  %s\n", bnNew.GetCompact(), bnNew.ToString());
-
-    return nNewBits;
-}
-
-bool CheckProofOfWork(uint256 hash, unsigned int nBits)
-{
-    bool fNegative;
-    bool fOverflow;
-    uint256 bnTarget;
-    bnTarget.SetCompact(nBits, &fNegative, &fOverflow);
-
-    // Check range
-    if (fNegative || bnTarget == 0 || fOverflow || bnTarget > Params().ProofOfWorkLimit())
-        return error("CheckProofOfWork() : nBits below minimum work");
-
-    // Check proof of work matches claimed amount
-    if (hash > bnTarget)
-        return error("CheckProofOfWork() : hash doesn't match nBits");
-
-    return true;
 }
 
 bool IsInitialBlockDownload()
